requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.1.0
soupsieve==2.6
typing_extensions==4.13.1
Werkzeug==3.1.3
